# contracts/openapi/runner-v1.yaml
openapi: 3.1.0
info:
  title: Runner Service API
  version: 1.0.0
  description: |
    The public, versioned contract of the Runner Service.

    The Runner is a standalone bounded service. It **executes** submitted Test
    IR; it does not author test cases and does not generate Test IR from
    natural language. A Test Authoring / IR Generation service integrates with
    the Runner through this document alone — there is no supported source-code
    dependency in either direction, and a caller may be written in any language.

    Three consumption styles are equally supported after submitting a run:
    poll `GET /api/v1/executions/{executionId}`, receive a webhook via
    `callback`, or consume the SSE stream at `/events`.
servers:
  - url: http://localhost:3001
    description: Local development
...
paths:
...
  /api/v1/executions:
    post:
      tags: [Executions]
      summary: Submit a Test IR execution
      description: |
        Validates the payload, stores an immutable snapshot of it for audit and
        replay, queues a browser job, and returns `202 Accepted` immediately.
        Browser work happens in a separate worker process.

        Send an `Idempotency-Key` header so a retried request returns the
        original execution instead of starting a second browser run.
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          schema: { type: string }
          description: Scoped per workspace. A repeat returns the original execution.
        - name: X-Request-Id
          in: header
          required: false
          schema: { type: string }
        - name: X-Correlation-Id
          in: header
          required: false
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExecutionRequestV1'
      responses:
        '202':
          description: The execution was accepted and queued.
          headers:
            Location:
              schema: { type: string }
              description: The status URL for this execution.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionAcceptedV1'
        '400':
          $ref: '#/components/responses/Error'
  /api/v1/executions/{executionId}:
    get:
      tags: [Executions]
      summary: Get execution status and result
      parameters:
        - $ref: '#/components/parameters/ExecutionId'
      responses:
        '200':
          description: The current state of the execution.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionResultV1'
        '404':
          $ref: '#/components/responses/Error'
  /api/v1/executions/{executionId}/cancel:
    post:
      tags: [Executions]
      summary: Request cancellation
      description: |
        Cancellation is cooperative. A queued job is removed; a running one
        stops at its next step boundary so the browser context and its
        artifacts are released cleanly.
      parameters:
        - $ref: '#/components/parameters/ExecutionId'
      responses:
        '202':
          description: Cancellation was requested.
          content:
            application/json:
              schema:
                type: object
                properties:
                  executionId: { type: string }
                  status: { type: string, example: CANCELLED }
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
...
  /api/v1/validate/execution:
    post:
      tags: [Validation]
      summary: Validate an execution request without running it
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExecutionRequestV1'
      responses:
        '200':
          description: The payload is valid.
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid: { type: boolean, example: true }
        '400':
          $ref: '#/components/responses/Error'
  /api/v1/validate/test-ir:
    post:
      tags: [Validation]
      summary: Validate a Test IR document without running it
      description: |
        Lets a generating service verify its output before scheduling a browser
        run, turning an execution-time failure into an authoring-time one.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TestIRV1'
      responses:
        '200':
          description: The Test IR is valid.
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid: { type: boolean, example: true }
        '400':
          $ref: '#/components/responses/Error'
...
components:
  parameters:
    ExecutionId:
      name: executionId
      in: path
      required: true
      schema: { type: string }
...
  responses:
    Error:
      description: A structured Runner error.
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                $ref: '#/components/schemas/RunnerError'
  schemas:
    RunnerError:
      type: object
      description: |
        `kind` separates a genuine test failure from a setup or infrastructure
        problem, so a reporting system never blames the application under test
        for a Runner-side issue.
      required: [code, kind, message, retryable]
      properties:
        code:
          type: string
          example: ELEMENT_NOT_FOUND
        kind:
          type: string
          enum:
            - TEST_FAILURE
            - PRECONDITION_FAILURE
            - RESOLUTION_FAILURE
            - INFRASTRUCTURE_FAILURE
            - CONTRACT_FAILURE
        message: { type: string }
        retryable: { type: boolean }
        details: { type: object }
    ElementIntentV1:
      type: object
      description: |
        How a step names its target. Note the absence of a raw selector:
        selectors live in the Runner Registry, so renaming an element or
        healing its selector never invalidates existing Test IR.
      minProperties: 1
      properties:
        elementId:
          type: string
          description: Stable Registry ID. The strongest form of reference.
        name:
          type: string
          example: Create Customer Button
        description:
          type: string
          example: Creates a new customer using values in the customer form.
        role: { type: string, example: button }
        semantic: { type: string }
        componentId: { type: string }
        pageId: { type: string }
    PreconditionV1:
      type: object
      description: |
        A requirement that must hold before the step runs. Failing to reach the
        state is reported as PRECONDITION_FAILED, never as a test failure.
      required: [type]
      properties:
        type:
          type: string
          enum: [authenticated, entityState, uiState, urlMatches, elementVisible]
        profile: { type: string, example: MANAGER }
        entity: { type: string, example: ORDER }
        state: { type: string, example: SUBMITTED }
        urlPattern: { type: string }
        target:
          $ref: '#/components/schemas/ElementIntentV1'
    AssertionV1:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum:
            [visible, hidden, text, containsText, value, enabled, disabled, checked, count, urlMatches]
        expected:
          oneOf:
            - { type: string }
            - { type: number }
            - { type: boolean }
        timeoutMs: { type: integer }
    TestActionV1:
      type: object
      required: [id, type]
      properties:
        id: { type: string }
        type:
          type: string
          enum: [goto, click, dblclick, fill, select, check, uncheck, hover, press, assert, wait]
        label: { type: string }
        target:
          $ref: '#/components/schemas/ElementIntentV1'
        value:
          oneOf:
            - { type: string }
            - { type: number }
            - { type: boolean }
        assertion:
          $ref: '#/components/schemas/AssertionV1'
        preconditions:
          type: array
          items:
            $ref: '#/components/schemas/PreconditionV1'
        timeoutMs: { type: integer }
        continueOnFailure: { type: boolean, default: false }
    TestIRV1:
      type: object
      required: [id, name, steps]
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string }
        steps:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/TestActionV1'
        metadata:
          type: object
          additionalProperties: { type: string }
    ExecutionRequestV1:
      type: object
      required: [contractVersion, irVersion, workspaceRef, test]
      properties:
        contractVersion:
          type: string
          enum: [runner.execution.v1]
        irVersion:
          type: string
          enum: [test-ir.v1]
        requestId: { type: string }
        tenantRef:
          type: string
          description: Opaque identifier owned by the caller. The Runner never interprets it.
        workspaceRef:
          type: string
          description: Scopes the Registry and every authorization check.
        externalTestCaseRef:
          type: string
          description: Opaque reference to the caller's test case, e.g. an Xray key.
        environmentRef: { type: string }
        authProfileRef:
          type: string
          description: Credentials are referenced by profile, never inlined.
        mode:
          type: string
          enum: [AUTO, REVIEW, INTERACTIVE]
          default: AUTO
        test:
          $ref: '#/components/schemas/TestIRV1'
        options:
          type: object
          properties:
            headless: { type: boolean, default: true }
            viewport:
              type: object
              properties:
                width: { type: integer }
                height: { type: integer }
            baseUrl: { type: string, format: uri }
            defaultTimeoutMs: { type: integer }
            stopOnFailure: { type: boolean, default: true }
            enableRegistryLearning: { type: boolean, default: true }
            enableSelfHealing: { type: boolean, default: false }
        callback:
          type: object
          required: [url, events]
          properties:
            url: { type: string, format: uri }
            events:
              type: array
              items:
                type: string
                enum:
                  - execution.queued
                  - execution.started
                  - execution.step.started
                  - execution.step.completed
                  - execution.waiting_user
                  - execution.completed
                  - execution.failed
                  - execution.cancelled
            signatureHeader: { type: string }
    ExecutionAcceptedV1:
      type: object
      required: [executionId, status, statusUrl, eventsUrl, acceptedAt]
      properties:
        executionId: { type: string, example: run_01j8x2c4 }
        status: { type: string, example: QUEUED }
        statusUrl: { type: string }
        eventsUrl: { type: string }
        requestId: { type: string }
        acceptedAt: { type: string, format: date-time }
    SelectorDefinition:
      type: object
      description: |
        A structured selector from the Runner selector DSL. Always data, never
        executable code, so a selector can be stored, diffed, scored and
        translated to another automation engine.
      required: [type]
      properties:
        type:
          type: string
          enum: [testId, role, label, placeholder, altText, title, text, css, xpath]
        value: { type: string }
        role: { type: string }
        name: { type: string }
        exact: { type: boolean }
    ExecutionStepResultV1:
      type: object
      required: [stepId, type, status]
      properties:
        stepId: { type: string }
        type: { type: string }
        status:
          type: string
          enum: [PENDING, RUNNING, PASSED, FAILED, SKIPPED]
        startedAt: { type: string, format: date-time }
        completedAt: { type: string, format: date-time }
        durationMs: { type: integer }
        resolvedElement:
          type: object
          properties:
            elementId: { type: string }
            displayName: { type: string }
            confidence: { type: number, minimum: 0, maximum: 1 }
            selector:
              $ref: '#/components/schemas/SelectorDefinition'
            alternatives:
              type: array
              items:
                $ref: '#/components/schemas/SelectorDefinition'
        evidence:
          type: array
          items: { type: string }
          description: Short human-readable reasons behind the resolution decision.
        error:
          $ref: '#/components/schemas/RunnerError'
        artifactIds:
          type: array
          items: { type: string }
    ExecutionResultV1:
      type: object
      required: [contractVersion, executionId, workspaceRef, status, mode, steps]
      properties:
        contractVersion: { type: string, enum: [runner.execution.v1] }
        executionId: { type: string }
        requestId: { type: string }
        workspaceRef: { type: string }
        externalTestCaseRef: { type: string }
        status:
          type: string
          enum: [QUEUED, RUNNING, WAITING_USER, PASSED, FAILED, CANCELLED]
        mode:
          type: string
          enum: [AUTO, REVIEW, INTERACTIVE]
        startedAt: { type: string, format: date-time }
        completedAt: { type: string, format: date-time }
        durationMs: { type: integer }
        steps:
          type: array
          items:
            $ref: '#/components/schemas/ExecutionStepResultV1'
        error:
          $ref: '#/components/schemas/RunnerError'
...